EWPP-376: Use DrushTestTrait and move test to BrowserTestBase.
