+++ composer.json
@@ -20,7 +20,7 @@
         "drupal/coder": "8.3.8",
         "drupal/config_devel": "~1.2",
         "drupal/drupal-extension": "~4.0",
-        "drush/drush": "~9.0@stable",
+        "drush/drush": "~9.7@stable",
         "guzzlehttp/guzzle": "~6.3",
         "openeuropa/code-review": "~1.0.0-beta2",
         "openeuropa/drupal-core-require-dev": "^8.7",
