+++ oe_contact_forms.services.yml
@@ -8,7 +8,7 @@ services:
     tags:
       - { name: access_check, applies_to: _oe_contact_forms_access_check }
   oe_contact_forms.contact.sanitize_commands:
-    class: Drupal\oe_contact_forms\Commands\sql\SanitizeContactFormFieldsCommands
-    arguments: ['@database', '@logger.factory', '@entity_type.manager', '@cache.entity']
+    class: Drupal\oe_contact_forms\Commands\sql\ContactFormSanitizeCommand
+    arguments: ['@entity_type.manager']
     tags:
       -  { name: drush.command }
